client // commands and instance
- Remove Args class
- remove manual construct of Process class
- Renamed Process to CommandRunner
- Run commands with Command.Ruh, Command.Watch
-

+++ src/MiniZinc.Parser/Syntax/FloatLiteralSyntax.cs
@@ -1,7 +1,7 @@
 ﻿namespace MiniZinc.Parser.Syntax;
 
 public sealed record FloatLiteralSyntax(in Token Start)
-    : SyntaxNode<decimal>(Start, Start.FloatValue)
+    : SyntaxNode<decimal>(Start, Start.DecimalValue)
 {
     public override string ToString() => Value.ToString("g");
 }
